+++ lua/render-markdown/request/highlights.lua
@@ -1,0 +1,147 @@
+local interval = require('render-markdown.lib.interval')
+
+---@class render.md.request.highlights.Line
+---@field hidden boolean
+---@field conceals render.md.request.highlights.Conceal[]
+
+---@class render.md.request.highlights.Entry
+---@field hidden? boolean
+---@field conceal? render.md.request.highlights.Conceal
+
+---@class render.md.request.highlights.Conceal: render.md.Range
+---@field [3] string replacement
+---@field [4] integer blocks
+
+---@class render.md.request.Highlights
+---@field private buf integer
+---@field private view render.md.request.View
+---@field private computed boolean
+---@field private lines table<integer, render.md.request.highlights.Line>
+local Highlights = {}
+Highlights.__index = Highlights
+
+---@param buf integer
+---@param view render.md.request.View
+---@return render.md.request.Highlights
+function Highlights.new(buf, view)
+    local self = setmetatable({}, Highlights)
+    self.buf = buf
+    self.view = view
+    self.computed = false
+    self.lines = {}
+    return self
+end
+
+---@param row integer
+---@param entry render.md.request.highlights.Entry
+function Highlights:add(row, entry)
+    if not self.lines[row] then
+        self.lines[row] = { hidden = false, conceals = {} }
+    end
+    local line = self.lines[row]
+    if entry.hidden then
+        line.hidden = entry.hidden
+    end
+    if entry.conceal then
+        if interval.valid(entry.conceal, true) then
+            line.conceals[#line.conceals + 1] = entry.conceal
+            line.conceals = Highlights.coalesce_conceals(line.conceals)
+        end
+    end
+end
+
+---@private
+---@param conceals render.md.request.highlights.Conceal[]
+---@return render.md.request.highlights.Conceal[]
+function Highlights.coalesce_conceals(conceals)
+    interval.sort(conceals)
+    local result = {} ---@type render.md.request.highlights.Conceal[]
+    result[#result + 1] = conceals[1]
+    for i = 2, #conceals do
+        local conceal, last = conceals[i], result[#result]
+        if conceal[1] <= last[2] then
+            last[2] = math.max(last[2], conceal[2])
+            last[3] = last[3] .. conceal[3]
+            last[4] = last[4] + conceal[4]
+        else
+            result[#result + 1] = conceal
+        end
+    end
+    return result
+end
+
+---@param body render.md.node.Body
+---@return render.md.request.highlights.Line
+function Highlights:line(body)
+    if not self.computed then
+        self.computed = true
+        self:compute()
+    end
+    local line = self.lines[body.start_row]
+    if not line then
+        line = { hidden = false, conceals = {} }
+    end
+    return line
+end
+
+---Cached row level implementation of vim.treesitter.get_captures_at_pos
+---@private
+function Highlights:compute()
+    if not vim.treesitter.highlighter.active[self.buf] then
+        return
+    end
+    local parser = vim.treesitter.get_parser(self.buf)
+    if not parser then
+        return
+    end
+    parser:for_each_tree(function(tree, language_tree)
+        self:tree(language_tree:lang(), tree:root())
+    end)
+end
+
+---@private
+---@param language string
+---@param root TSNode
+function Highlights:tree(language, root)
+    if not self.view:overlaps(root) then
+        return
+    end
+    if not vim.tbl_contains({ 'markdown', 'markdown_inline' }, language) then
+        return
+    end
+    local query = vim.treesitter.query.get(language, 'highlights')
+    if not query then
+        return
+    end
+    self.view:query(root, query, function(id, node, data)
+        if data.conceal_lines then
+            local row = Highlights.range(id, data, node)
+            self:add(row, { hidden = true })
+        end
+        if data.conceal then
+            local row, start_col, _, end_col = Highlights.range(id, data, node)
+            self:add(row, {
+                conceal = { start_col, end_col, data.conceal, 1 },
+            })
+        end
+    end)
+end
+
+---@private
+---@param id integer
+---@param data vim.treesitter.query.TSMetadata
+---@param node TSNode
+---@return integer, integer, integer, integer
+function Highlights.range(id, data, node)
+    local range = (data[id] or {}).range
+    if range then
+        return range[1], range[2], range[3], range[4]
+    end
+    range = data.range
+    if range then
+        return range[1], range[2], range[3], range[4]
+    end
+    return node:range()
+end
+
+return Highlights