+++ lua/render-markdown/health.lua
@@ -13,7 +13,7 @@ local state = require('render-markdown.state')
 local M = {}
 
 ---@private
-M.version = '8.13.8'
+M.version = '8.14.0'
 
 function M.check()
     M.start('versions')